Renderer2D + maintenance

+++ Engine/src/Engine/Renderer/Buffer.cpp
@@ -42,7 +42,7 @@ namespace Engine {
 	}
 
 	void BufferLayout::calculateOffsetsAndStride() {
-		unsigned int offset = 0;
+		size_t offset = 0;
 		m_stride = 0;
 
 		for (auto& element : m_elements) {
